requirements.txt
# Document Q&A System Dependencies

## Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
customtkinter>=5.0.0

## Optional Dependencies (for development)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
